Use the same signature in 'fields_for'
Fixes one of the allowed parameter combinations for `fields_for` method. The
combination that this fixes is when you pass the attribute name, and the
object, ie:

```ruby
fields_for(:attribute, nested_objecta) {|f| ...}
```

Before this fix the method renders nothing as the `record_object` as defined in
the original Rails implementation gets lost in the gem version. I've made the
method signature the same as the Rails version while preserving the previous
functionality of setting the builder in the options.