Failing test case for nested buiders nil error
When there is a nested builder that renders errors before the parent builder the
parent builder raises a nil error.

I believe the issue is because the initializer is not thread safe since it
overrides `ActionView::Base.field_error_proc` with a proc that is coupled with
form resource.

I think this might be a much bigger issue for applications served with threaded
servers such as Puma because requests from different people will override
`field_error_proc` with potentially strange results such as rendering forms from
different people or erroring randomly. Since Rails 5 Puma is also the default
webserver.