Add error reporting when generating fieldsets
Styling for errors has been taken from the govuk elements examples here:

https://govuk-elements.herokuapp.com/form-elements/example-form-elements/

Some of the example HTML on the above page is inconsistent with HTML
given as examples here:

https://govuk-elements.herokuapp.com/form-elements/#form-radio-buttons

We've aligned with the HTML from the former link because it seems to
better support errors. This has meant tweaking some of the non-error
HTML, like adding <span> tags around tags in legends.

Also fixed some of the HTML generated so that it doesn't get mangled to
be "HTML-safe".