Allow specifying a CSS class for fieldset legend
A common pattern for simple question pages with a single fieldset is
to hide the legend using CSS so as to avoid duplication (the question
will already have been asked in an <hX> tag) but still allow
screenreaders to access it.

This change allows the user to specify an HTML class attribute on the
legend to do this (or indeed something else).