Add ability to specify HTML safe label translations
It's a common pattern on some applications to use (limited) HTML
markup inside labels. Rails has a pattern where it assumes that
translations are html_safe when they end in `_html`. This PR adds
the ability to add html_safe labels to forms by having the
translation code look for `fieldname_html` in the translations
before falling back to the default, and if it finds that,
assuming the resulting translation is html_safe.